Updata Project For alert when User Delete Acout alert to Admin

+++ database/migrations/2026_03_20_161914_add_name_to_order_items_table.php
@@ -18,7 +18,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('order_items', function (Blueprint $table) {
-            $table->dropColumn('name');
+            if (Schema::hasColumn('order_items', 'name')) {
+                $table->dropColumn('name');
+            }
         });
     }
 };
